backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "tracker backend"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "pyjwt[crypto]>=2.10.1",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]

[tool.ruff]
indent-width = 2

[tool.ruff.lint]
ignore = ["E305"]
preview = true
